update to chromium 86.0.4240.111

+++ components/printing/renderer/print_render_frame_helper.cc
@@ -2078,8 +2078,10 @@ void PrintRenderFrameHelper::IPCReceived() {
 
 void PrintRenderFrameHelper::IPCProcessed() {
   --ipc_nesting_level_;
-  if (ipc_nesting_level_ == 0 && render_frame_gone_)
+  if (ipc_nesting_level_ == 0 && render_frame_gone_ && !delete_pending_) {
+    delete_pending_ = true;
     base::ThreadTaskRunnerHandle::Get()->DeleteSoon(FROM_HERE, this);
+  }
 }
 
 bool PrintRenderFrameHelper::InitPrintSettings(bool fit_to_paper_size) {
@@ -2471,6 +2473,10 @@ bool PrintRenderFrameHelper::PreviewPageRendered(
                "page_number", page_number);
 
 #if BUILDFLAG(ENABLE_TAGGED_PDF)
+  // Make sure the RenderFrame is alive before taking the snapshot.
+  if (render_frame_gone_)
+    snapshotter_.reset();
+
   // For tagged PDF exporting, send a snapshot of the accessibility tree
   // along with page 0. The accessibility tree contains the content for
   // all of the pages of the main frame.
